bot/requirements.txt
aiogram==2.23.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
SQLAlchemy
psycopg2-binary